requirements.txt
gradio>=4.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pypdf>=3.0.0
huggingface_hub>=0.20.0
transformers>=4.35.0
torch>=2.0.0
